v0.3.1

+++ run.sh
@@ -1,9 +1,13 @@
 #!/bin/bash
+make clean
+if make; then
+    dd if=/dev/zero of=disk.img bs=1M count=10
+    mkfs.ext2 -F disk.img
+    debugfs -w -R "write hello.bin hello.bin" disk.img
+    qemu-system-i386 -kernel pixelos.bin \
+                 -drive file=disk.img,format=raw,index=0,media=disk \
+                 -d int -no-reboot
 
-# Создаем пустой файл размером 10 МБ
-dd if=/dev/zero of=disk.img bs=1M count=10
-
-# Форматируем его в ext2 (ответь 'y', если спросит про файл)
-mkfs.ext2 disk.img
-
-qemu-system-i386 -kernel pixelos.bin -hda disk.img
+else
+    echo "Build failed, check your code!"
+fi
